+++ db/updates/update_13.4-13.5.sql
@@ -1,0 +1,1 @@
+DELETE FROM `system_settings` WHERE `system_settings`.`key_name` = 'LIBRARY_apiurl'