+++ pages/ingredients.php
@@ -204,7 +204,6 @@ $(document).ready(function() {
 	});
 	
 	list_ingredients();
-	var pvLibraryAPI = '<?php echo $pvLibraryAPI; ?>';
 	$(function () {
 		$(".input-group-btn .dropdown-menu li a").click(function () {
 			var selText = $(this).html();